refactor(auv-cli): extract inspect donor render modules (4B)
Move Minecraft/Osu/Balatro text render sections out of the inspect facade into inspect/{minecraft,osu,balatro}.rs while keeping inspect_run orchestration in the root module.



effort: max
